move the sightinglist JS
