+++ server/scripts/sightinglist.js
